Refactor a bit into two classes
